refactor: use hooks to handle store and remove useEffect to format data in DomainFilter
