+++ src/App.tsx
@@ -1,16 +1,16 @@
 // This file has to be left untouched
 
-import React, { Component } from 'react';
-import DomainFilter from './components/DomainFilter';
+import React, { Component } from 'react'
+import DomainFilter from './components/DomainFilter'
 
 class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <DomainFilter />
-      </div>
-    );
-  }
+	render() {
+		return (
+			<div className="App">
+				<DomainFilter />
+			</div>
+		)
+	}
 }
 
-export default App;
+export default App
