+++ src/components/DomainFilter/DomainFilter.test.tsx
@@ -1,18 +1,17 @@
-import DomainFilter from './DomainFilter.component';
+import DomainFilter from './DomainFilter';
 import { render, screen } from '@testing-library/react';
 
-describe('components', () => {
-  describe('DomainFilter', () => {
-    it('should allow the user to filter', () => {
-      render(<DomainFilter domains={['do']} />);
+// TODO fix tests
+describe('DomainFilter', () => {
+  it('should allow the user to filter', () => {
+    render(<DomainFilter />);
 
-      expect(screen.getAllByRole('listbox')).toHaveLength(3);
-    })
-
-    it('should render', async () => {
-      render(<DomainFilter domains={['do']} />);
-
-      expect(await screen.findByText('do')).toBeTruthy();
-    });
+    expect(screen.getAllByRole('listbox')).toHaveLength(3);
   })
+
+  it('should render', async () => {
+    render(<DomainFilter />);
+
+    expect(await screen.findByText('do')).toBeTruthy();
+  });
 })
