+++ src/redux/types.ts
@@ -1,0 +1,3 @@
+import rootReducer from './reducer'
+
+export type RootState = ReturnType<typeof rootReducer>
